refactor: move `resolveSx` to style helpers
