+++ src/shared/lib/helpers/styles/styles.ts
@@ -8,3 +8,14 @@ export type SX = SxProps<Theme>;
 export function bgIconCenter(icon: string) {
   return { background: `url(${icon}) no-repeat center` };
 }
+/**
+ * Объединяет `sx`-стили. (Не работает с массивами)
+ */
+export function resolveSx<T extends Theme>(
+  ...sxs: Array<SxProps<T> | undefined>
+): SxProps<T> {
+  return sxs.reduce((resolvedSx, sx): SxProps<T> => {
+    if (sx === undefined) sx = {};
+    return Object.assign(resolvedSx as object, sx);
+  }, {}) as SxProps<T>;
+}
